refactor(forgesvc): move forward+reverse board-sync seam to projectsync_sync.go (#963)
Pure same-package code motion, move-only (PRD #963 M2). Moves the forward and
reverse board-sync surface — ForwardMove, addForwardItem, ReverseSync,
reconcileItems, backfillItem, the plannedMove type, and reverseDiff — plus the
three sync-only helpers stampLinkErrorReverse, stampLinkError and markerValue,
verbatim (bodies and doc comments byte-identical) out of projectsync.go into a
new sibling file projectsync_sync.go, same package forgesvc.

Caller map re-derived at this base: those three helpers are still called only
from within the sync-seam functions. ProjectSyncStatus, ensureProjectScope,
optionMarker and truncateErr stay in projectsync.go. No exported-API change, no
behavior change, no test edit; imports settled by the compiler. Verified
byte-identical motion (764 moved lines diff-clean) and go build/vet green.
